change level class to float, add conversion callback, create default system data

+++ db/migrate/20220422235805_create_alerts.rb
@@ -2,7 +2,7 @@ class CreateAlerts < ActiveRecord::Migration[6.1]
   def change
     create_table :alerts do |t|
       t.references :user, null: false, foreign_key: true
-      t.integer :level
+      t.float :level
 
       t.timestamps
     end
