+++ db/migrate/20220423222014_create_seconds.rb
@@ -2,7 +2,7 @@ class CreateSeconds < ActiveRecord::Migration[6.1]
   def change
     create_table :seconds do |t|
       t.string :title
-      t.integer :level
+      t.float :level
       t.references :user, null: false, foreign_key: true
 
       t.timestamps
